feat: 识别文件类型：音乐、电影、电视剧
