+++ app/modules/musicbrainz/__init__.py
@@ -4,14 +4,14 @@ from typing import Any, Optional, Tuple, Union
 
 from app.core.cache import cached
 from app.core.config import settings
-from app.core.meta import MetaMusic
-from app.core.music import (
+from app.core.context import (
     MUSIC_ENTITY_ALBUM,
     MusicAlbumInfo,
     MusicArtistInfo,
     MusicInfo,
     MusicRelease,
 )
+from app.core.meta import MetaMusic
 from app.log import logger
 from app.modules import _ModuleBase
 from app.schemas.types import MediaRecognizeType, ModuleType
