+++ tests/test_media_scrape_endpoint.py
@@ -3,7 +3,7 @@ from unittest.mock import AsyncMock, Mock, patch
 from app.api.endpoints.media import recognize_file, scrape
 from app.core.context import Context, MediaInfo
 from app.core.meta import MetaBase, MetaMusic
-from app.core.music import MusicInfo
+from app.core.context import MusicInfo
 from app.schemas import FileItem, MediaType
