+++ tests/test_music_search.py
@@ -2,7 +2,7 @@ from unittest.mock import Mock, patch
 
 from app.chain.search import SearchChain
 from app.core.meta import MetaMusic
-from app.core.music import MusicInfo
+from app.core.context import MusicInfo
 from app.schemas.context import TorrentInfo
 from app.schemas.types import MediaType
